# src/config/openapi.yaml
openapi: 3.0.3
info:
  title: Time Recording System API
  version: 1.1.0
  description: |
    A RESTful API for recording employee work hours, managing work calendars, and generating reports.

    ## Authentication
    All endpoints except `POST /auth/login` and `GET /health` require a valid JWT in the
    `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    Obtain a token by calling `POST /auth/login` with valid credentials.

    ## Response Envelope
    Every response is wrapped in a consistent envelope:
    ```json
    { "success": true,  "data": { ... } }
    { "success": false, "error": "Human-readable error message" }
    ```

    ## Concurrency
    Clock-in and clock-out operations run inside **serializable PostgreSQL transactions**,
    ensuring correctness under concurrent requests. A concurrent duplicate clock-in will
    be rejected with HTTP `409 Conflict`.

    ## Rate Limiting
    | Scope | Limit | Window |
    |---|---|---|
    | `POST /auth/login` | 10 requests (failed only) | 15 minutes |
    | All other routes | 100 requests | 1 minute |
    Exceeding either limit returns HTTP `429 Too Many Requests`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server

# Global default: all endpoints require BearerAuth unless overridden with security: []
security:
  - BearerAuth: []
...
components:
  schemas:
...
    Uuid:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    IsoDateTime:
      type: string
      format: date-time
      example: "2024-03-15T09:00:00.000Z"
    DateString:
      type: string
      format: date
      example: "2024-03-15"

    # ── Envelope ──────────────────────────────────────────────────────────────
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "User is already clocked in"

    # ── User ──────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
          example: "Alice Smith"
        email:
          type: string
          format: email
          example: "alice@example.com"
        role:
          type: string
          enum: [employee, admin]
          description: "employee = regular staff; admin = can modify work config and calendar overrides"
          example: "employee"
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
        updatedAt:
          $ref: '#/components/schemas/IsoDateTime'
    UserSummary:
      type: object
      required: [id, name, email]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
          example: "Alice Smith"
        email:
          type: string
          format: email
          example: "alice@example.com"
    CreateUserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          example: "Alice Smith"
        email:
          type: string
          format: email
          example: "alice@example.com"
        password:
          type: string
          minLength: 6
          example: "securePass123"
        role:
          type: string
          enum: [employee, admin]
          default: employee
          example: "employee"
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          example: "Alice Johnson"
        email:
          type: string
          format: email
          example: "alice.johnson@example.com"

    # ── Time Record ───────────────────────────────────────────────────────────
    TimeRecord:
      type: object
      required: [id, userId, clockIn, date, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        userId:
          $ref: '#/components/schemas/Uuid'
        user:
          $ref: '#/components/schemas/UserSummary'
        clockIn:
          $ref: '#/components/schemas/IsoDateTime'
        clockOut:
          $ref: '#/components/schemas/IsoDateTime'
          nullable: true
        date:
          $ref: '#/components/schemas/IsoDateTime'
          description: Midnight UTC of the local working date (derived from clockIn + APP_TIMEZONE)
        workedMinutes:
          type: integer
          nullable: true
          description: Total minutes worked. Null while session is open.
          example: 570
        overtimeMinutes:
          type: integer
          nullable: true
          description: Minutes exceeding normalHoursPerDay. Null while session is open.
          example: 90
        notes:
          type: string
          nullable: true
          example: "Sprint planning day"
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
        updatedAt:
          $ref: '#/components/schemas/IsoDateTime'
    CreateTimeRecordRequest:
      type: object
      required: [userId, clockIn]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
        clockIn:
          $ref: '#/components/schemas/IsoDateTime'
        clockOut:
          $ref: '#/components/schemas/IsoDateTime'
          description: Optional. If omitted the record is left open (session in progress).
        notes:
          type: string
          example: "Manually entered record"
    UpdateTimeRecordRequest:
      type: object
      properties:
        clockIn:
          $ref: '#/components/schemas/IsoDateTime'
        clockOut:
          $ref: '#/components/schemas/IsoDateTime'
          nullable: true
          description: Set to null to re-open a completed record.
        notes:
          type: string
          nullable: true
    PaginatedTimeRecords:
      type: object
      required: [records, total, page, limit, totalPages]
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/TimeRecord'
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 3

    # ── Clock ─────────────────────────────────────────────────────────────────
    ClockRequest:
      type: object
      required: [userId]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
        notes:
          type: string
          example: "Starting deep-focus session"
    ClockStatus:
      type: object
      required: [userId, isClockedIn]
      properties:
        userId:
          $ref: '#/components/schemas/Uuid'
        isClockedIn:
          type: boolean
          example: true
        currentRecord:
          nullable: true
          type: object
          description: Present only when isClockedIn is true.
          required: [id, clockIn]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            clockIn:
              $ref: '#/components/schemas/IsoDateTime'
            notes:
              type: string
              nullable: true

    # ── Work Config ───────────────────────────────────────────────────────────
    WorkConfig:
      type: object
      required: [id, normalHoursPerDay, workingDaysOfWeek, effectiveFrom]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        normalHoursPerDay:
          type: integer
          minimum: 1
          maximum: 24
          example: 8
        workingDaysOfWeek:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          description: "Day-of-week numbers: 0=Sun, 1=Mon, …, 6=Sat"
          example: [1, 2, 3, 4, 5]
        effectiveFrom:
          $ref: '#/components/schemas/IsoDateTime'
        createdAt:
          $ref: '#/components/schemas/IsoDateTime'
        updatedAt:
          $ref: '#/components/schemas/IsoDateTime'
    WorkConfigRequest:
      type: object
      required: [normalHoursPerDay, workingDaysOfWeek]
      properties:
        normalHoursPerDay:
          type: integer
          minimum: 1
          maximum: 24
          example: 8
        workingDaysOfWeek:
          type: array
          minItems: 1
          items:
            type: integer
            minimum: 0
            maximum: 6
          description: "Day-of-week numbers: 0=Sun, 1=Mon, …, 6=Sat"
          example: [1, 2, 3, 4, 5]

    # ── Calendar Override ─────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error: "Missing or malformed Authorization header"
    Forbidden:
      description: Authenticated but insufficient role (admin required)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error: "Admin access required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error: "User not found"
    BadRequest:
      description: Validation error or invalid business-rule transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error: "User is already clocked in"
    Conflict:
      description: Concurrent request conflict — retry the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error: "Request conflict, please try again"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `POST /auth/login`.
        Include in every protected request:
        ```
        Authorization: Bearer <token>
        ```
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              name: "Alice Smith"
              email: "alice@example.com"
              password: "securePass123"
              role: "employee"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Users]
      summary: List all users
      operationId: listUsers
      responses:
        "200":
          description: Array of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [Users]
      summary: Get a user by ID
      operationId: getUser
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: Update a user's name or email
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Delete a user (cascades to their time records)
      operationId: deleteUser
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                message: "User deleted"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Clock ─────────────────────────────────────────────────────────────────
  /clock/in:
    post:
      tags: [Clock]
      summary: Clock a user in
      description: |
        Starts a new work session for the user. Fails with **400** if the user
        already has an open (not yet clocked-out) session.

        Operations run inside a **serializable transaction** to prevent concurrent
        duplicate clock-ins. If two requests race, one will succeed (201) and the
        other will be rejected — either with **400** (caught by app logic) or
        **409** (caught by the database serialization guard).
      operationId: clockIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockRequest'
            example:
              userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
              notes: "Starting sprint planning"
      responses:
        "201":
          description: Clocked in successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeRecord'
              example:
                success: true
                message: "Clocked in successfully"
                data:
                  id: "a1b2c3d4-0000-0000-0000-000000000001"
                  userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  clockIn: "2024-03-15T09:00:00.000Z"
                  clockOut: null
                  date: "2024-03-15T00:00:00.000Z"
                  workedMinutes: null
                  overtimeMinutes: null
                  notes: "Starting sprint planning"
        "400":
          description: User already clocked in or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error: "User is already clocked in (record a1b2c3d4-..., since 2024-03-15T08:00:00.000Z)"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
  /clock/out:
    post:
      tags: [Clock]
      summary: Clock a user out
      description: |
        Closes the user's current open session. Automatically computes
        `workedMinutes` and `overtimeMinutes` based on the current work config.

        Fails with **400** if the user has no open session.
      operationId: clockOut
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockRequest'
            example:
              userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      responses:
        "200":
          description: Clocked out successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeRecord'
              example:
                success: true
                message: "Clocked out successfully"
                data:
                  id: "a1b2c3d4-0000-0000-0000-000000000001"
                  userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  clockIn: "2024-03-15T09:00:00.000Z"
                  clockOut: "2024-03-15T18:30:00.000Z"
                  date: "2024-03-15T00:00:00.000Z"
                  workedMinutes: 570
                  overtimeMinutes: 90
                  notes: null
        "400":
          description: User is not clocked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error: "User is not clocked in"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
  /clock/status/{userId}:
    get:
      tags: [Clock]
      summary: Get a user's current clock status
      operationId: getClockStatus
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        "200":
          description: Current clock status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ClockStatus'
              examples:
                clocked_in:
                  summary: User is clocked in
                  value:
                    success: true
                    data:
                      userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      isClockedIn: true
                      currentRecord:
                        id: "a1b2c3d4-0000-0000-0000-000000000001"
                        clockIn: "2024-03-15T09:00:00.000Z"
                        notes: null
                clocked_out:
                  summary: User is not clocked in
                  value:
                    success: true
                    data:
                      userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      isClockedIn: false
                      currentRecord: null
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Time Records ──────────────────────────────────────────────────────────
  /time-records:
    post:
      tags: [Time Records]
      summary: Manually create a time record
      description: |
        Allows back-filling or manually entering a complete clock-in/out record.
        `workedMinutes` and `overtimeMinutes` are auto-computed if `clockOut` is provided.
      operationId: createTimeRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeRecordRequest'
            example:
              userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
              clockIn: "2024-03-14T09:00:00Z"
              clockOut: "2024-03-14T18:30:00Z"
              notes: "Backfilled record"
      responses:
        "201":
          description: Time record created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeRecord'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Time Records]
      summary: List time records
      description: Returns a paginated list of records. Filter by user and/or date range.
      operationId: listTimeRecords
      parameters:
        - name: userId
          in: query
          schema:
            $ref: '#/components/schemas/Uuid'
          description: Filter by user ID
        - name: from
          in: query
          schema:
            $ref: '#/components/schemas/DateString'
          description: Start date (inclusive), format YYYY-MM-DD
        - name: to
          in: query
          schema:
            $ref: '#/components/schemas/DateString'
          description: End date (inclusive), format YYYY-MM-DD
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated records
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedTimeRecords'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /time-records/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [Time Records]
      summary: Get a time record by ID
      operationId: getTimeRecord
      responses:
        "200":
          description: Time record found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeRecord'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Time Records]
      summary: Update a time record
      description: |
        Partial update — only supply the fields you want to change.
        `workedMinutes` and `overtimeMinutes` are recomputed automatically.
        Set `clockOut` to `null` to re-open a closed record.
      operationId: updateTimeRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTimeRecordRequest'
            example:
              clockOut: "2024-03-14T17:00:00Z"
              notes: "Left early"
      responses:
        "200":
          description: Time record updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TimeRecord'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Time Records]
      summary: Delete a time record
      operationId: deleteTimeRecord
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                message: "Time record deleted"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Reports ───────────────────────────────────────────────────────────────
...
  /config:
    get:
      tags: [Work Config]
      summary: Get the current work configuration
      operationId: getWorkConfig
      responses:
        "200":
          description: Current work config
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkConfig'
              example:
                success: true
                data:
                  id: "9c1f5a00-0000-0000-0000-000000000001"
                  normalHoursPerDay: 8
                  workingDaysOfWeek: [1, 2, 3, 4, 5]
                  effectiveFrom: "2024-01-01T00:00:00.000Z"
        "401":
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Work Config]
      summary: Update the work configuration (admin only)
      description: |
        Overwrites the single active configuration row.
        Changes take effect immediately for all future overtime calculations.

        **Requires admin role.**
      operationId: updateWorkConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkConfigRequest'
            examples:
              standard:
                summary: Standard 5-day, 8 h/day
                value:
                  normalHoursPerDay: 8
                  workingDaysOfWeek: [1, 2, 3, 4, 5]
              four_day_week:
                summary: 4-day work week, 9 h/day
                value:
                  normalHoursPerDay: 9
                  workingDaysOfWeek: [1, 2, 3, 4]
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WorkConfig'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'

  # ── Calendar Overrides ────────────────────────────────────────────────────
...